GDQ plugin beta
This plugin lets you get all sorts of information on GDQ, such as links, the current donation total, and what games are in the schedule. There's also a basic notification command that alerts you when your favorite game is being played.

It doesn't use the best notification system ever, but it's something.