Chat mirroring, track limits, and more
The player can now mirror the chat to another satellite message between the main interface and the album art. Also, past notifications are logged in this message. This keeps the player at the bottom of the channel.

There can now be track limits on both the entire tracklist and individual users.

Permissions have changed slightly. In partially public mode, users can now start and pause the player, and skip tracks that they've added themselves.